Use wrappers around some of the uses of raw pointers
For `Use`s, when the actual use is dropped any instructions that
reference it have the appropriate operand nulled out. This wraps the
pointer in one that allows for easier usage of them in safe code.

Similarly, since types are valid for as long as the context exists, wrap
them in a type that derefs to the appropriate type.